refactor: extract self_update and preflight into sibling modules to satisfy structure check
- crates/cp-orchestrator/src/services/releases/mod.rs (537->416 lines): move the
  orchestrator self-update section (stage_orchestrator_update, boot_check,
  boot_commit, backup_path, pending_path, with_suffix, MAX_BOOT_ATTEMPTS) into a
  new sibling self_update.rs; re-export the public fns so external paths
  (crate::services::releases::{stage_orchestrator_update,boot_check,boot_commit})
  and the test helpers (backup_path, pending_path) are unchanged.
- src/modules/overview/mod.rs (501->~412 lines): move the Module::pre_flight body
  into tools/preflight.rs and delegate. Placed inside the existing tools/ subdir
  (not the overview/ root) so overview/ stays at 8 directory entries (max 8).

No behaviour change. check-structure --ci passes (lengths <=500, entries <=8,
chain intact); full rust lint gate green (fmt, clippy -D warnings, check,
lint-exceptions, vault-bypass).
