+++ src/modules/overview/tools/mod.rs
@@ -4,3 +4,5 @@ pub(super) mod close_context;
 pub(super) mod manage_tools;
 /// Tool for navigating paginated panels.
 pub(super) mod panel_goto_page;
+/// Pre-flight validation for the core module's tools.
+pub(super) mod preflight;
